Refactor PublicGallery component
Breaks down the large PublicGallery component into smaller, more manageable components for improved readability and maintainability.
